requirements.txt
asgiref==3.5.1
certifi==2021.10.8
charset-normalizer==2.0.12
decorator==5.1.1
Django==4.0.4
django-crispy-forms==1.14.0
django-environ==0.8.1
django-jalali-date==1.0.1
django-tempus-dominus==5.1.2.17
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
environ==1.0
et-xmlfile==1.1.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
jdatetime==4.1.0
numpy==1.22.3
openpyxl==3.0.10
pandas==1.4.2
psycopg2==2.9.3
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
self==2020.12.3
six==1.16.0
SQLAlchemy==1.4.36
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
xlrd==2.0.1
xlwt==1.3.0

